TestShell/TestLogger.cpp 추가 & log/latest.log 출력
