+++ TestShell/TestShell.vcxproj.filters
@@ -16,10 +16,10 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-    <None Include="run_list.lst">
+    <None Include="SSD.exe">
       <Filter>리소스 파일</Filter>
     </None>
-    <None Include="SSD.exe">
+    <None Include="run_list.lst">
       <Filter>리소스 파일</Filter>
     </None>
   </ItemGroup>
@@ -43,6 +43,9 @@
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gmock\gmock-all.cc">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="TestLogger.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
